adding a few fixes
