+++ PWGJE/DataModel/Jet.h
@@ -202,9 +202,9 @@ using CandidatesLcData = o2::soa::Join<o2::aod::Hf3PBases, o2::aod::Hf3PPars, o2
 using CandidatesLcMCD = o2::soa::Join<o2::aod::Hf3PBases, o2::aod::Hf3PPars, o2::aod::Hf3PParEs, o2::aod::Hf3PSels, o2::aod::Hf3PMls, o2::aod::Hf3PMcs, o2::aod::JLcIds>;
 using JetTracksSubLc = o2::aod::JTrackLcSubs;
 
-using CandidatesV0Data = o2::soa::Join<o2::aod::V0Cores, o2::aod::V0Extras, o2::aod::JV0Ids>;
-using CandidatesV0MCD = o2::soa::Join<o2::aod::V0Cores, o2::aod::V0Extras, o2::aod::V0MCCores, o2::aod::JV0Ids, o2::aod::JV0Ids>; // add a table for McV0Labels with daughter prongs as well
-using V0Daughters = o2::aod::DauTrackExtras;                                                                                      // linked by  V0Extras - check what this is
-using CandidatesV0MCP = o2::soa::Join<o2::aod::JV0McParticles, o2::aod::JV0PIds>;
+using CandidatesV0Data = o2::soa::Join<o2::aod::V0Cores, o2::aod::JV0Ids>;
+using CandidatesV0MCD = o2::soa::Join<o2::aod::V0Cores, o2::aod::V0MCCores, o2::aod::JV0Ids>;
+// using V0Daughters = o2::aod::DauTrackExtras;                                                                                      // we need to add the V0Extras table to CandiatesV0Data if we want to use this. Will wait for an analysis use case
+using CandidatesV0MCP = o2::aod::JV0McParticles;
 
 #endif // PWGJE_DATAMODEL_JET_H_
